2006-06-26  Andrew John Hughes  <gnu_andrew@member.fsf.org>
	* gnu/java/lang/management/ClassLoadingMXBeanImpl.java,
	* java/lang/management/ClassLoadingMXBean.java:
	(getTotalLoadedClassCount()): Corrected return type.
	(getUnloadedClassCount()): Likewise.
	* vm/reference/gnu/java/lang/management/VMClassLoadingMXBeanImpl.java:
	(getUnloadedClassCount()): Likewise.

+++ gnu/java/lang/management/ClassLoadingMXBeanImpl.java
@@ -58,12 +58,12 @@ public final class ClassLoadingMXBeanImpl
     return VMClassLoadingMXBeanImpl.getLoadedClassCount();
   }
 
-  public int getTotalLoadedClassCount()
+  public long getTotalLoadedClassCount()
   {
     return getLoadedClassCount() + getUnloadedClassCount();
   }
 
-  public int getUnloadedClassCount()
+  public long getUnloadedClassCount()
   {
     return VMClassLoadingMXBeanImpl.getUnloadedClassCount();
   }
